Realistic Calendar Dates
We now calculate the correct day of the week for our first date!

This change will also dynamically make the weeks instead of hardcoding. However updates will be made in the next commit to fix an issue with the last date stretching (since our ui depends on 7blocks for sizing)